One_Time_Multi_Game_Setup_Script
As the title suggests this is a script with the goal of setting up all supported intalled and managed games with a single click/command. The note in the README about symlinks inspired this. Not sure how this will be recieved but so far it seems to work well with Fallout New Vegas